Clean up the tests
