+++ juniper-eager-loading/tests/integration_tests.rs
@@ -1,4 +1,9 @@
+#![allow(unused_variables, unused_imports, dead_code, unused_mut)]
+
+mod helpers;
+
 use assert_json_diff::{assert_json_eq, assert_json_include};
+use helpers::{SortedExtension, StatsHash};
 use juniper::{Executor, FieldError, FieldResult};
 use juniper_eager_loading::{
     prelude::*, EagerLoading, HasMany, HasManyThrough, HasOne, OptionHasOne,
@@ -1168,68 +1173,4 @@ fn run_query(query: &str, db: Db) -> (Value, DbStats) {
             employment_reads: ctx.db.employments.reads_count(),
         },
     )
-}
-
-struct StatsHash<K: Hash + Eq, V> {
-    map: HashMap<K, V>,
-    count: AtomicUsize,
-    name: &'static str,
-}
-
-impl<K: Hash + Eq, V> StatsHash<K, V> {
-    fn new(name: &'static str) -> Self {
-        StatsHash {
-            map: HashMap::default(),
-            count: AtomicUsize::default(),
-            name,
-        }
-    }
-
-    #[allow(dead_code)]
-    fn get<Q>(&self, k: &Q) -> Option<&V>
-    where
-        K: Borrow<Q>,
-        Q: ?Sized + Hash + Eq,
-    {
-        self.increment_reads_count();
-        self.map.get(k)
-    }
-
-    #[allow(dead_code)]
-    pub fn get_mut<Q: ?Sized>(&mut self, k: &Q) -> Option<&mut V>
-    where
-        K: Borrow<Q>,
-        Q: Hash + Eq,
-    {
-        self.increment_reads_count();
-        self.map.get_mut(k)
-    }
-
-    fn all_values(&self) -> Vec<&V> {
-        self.increment_reads_count();
-        self.map.iter().map(|(_, v)| v).collect()
-    }
-
-    fn reads_count(&self) -> usize {
-        self.count.load(Ordering::SeqCst)
-    }
-
-    fn insert(&mut self, k: K, v: V) -> Option<V> {
-        self.map.insert(k, v)
-    }
-
-    fn increment_reads_count(&self) {
-        self.count.fetch_add(1, Ordering::SeqCst);
-    }
-}
-
-trait SortedExtension {
-    fn sorted(self) -> Self;
-}
-
-impl<T: std::cmp::Ord> SortedExtension for Vec<T> {
-    fn sorted(mut self) -> Self {
-        self.sort();
-        self
-    }
 }
